Add two Neovim plugins to editor support chapter
I changed the name of the section; since both editors have a lot in
common it is not necessary to have two sections, but since Neovim has
exclusive plugins it warrants being mentioned explicitly. If there are
more exclusive packages in the future we might consider making Neovim a
subsection of Vim.