Don't wrap SWANK:CREATE-SERVER inside BT:MAKE-THREAD
In general, a call to SWANK:CREATE-SERVER is not blocking (i.e. it
creates a thread on its own), hence wrapping it inside BT:MAKE-THREAD is
kind of redundant.  The only situations for which a call to
BT:MAKE-THREAD would be required are as if:

- the underlying implementation does not support multi-threading, in
which case, a call to BT:MAKE-THREAD would fail anyway
- the user is explicitly forcing a _blocking_ communication style (i.e.
`(setf swank:*communication-style* nil)`), but since we, as the users,
are not enforcing that, then we should be good without that extra call

While at it, I also:

- made QL:QUICKLOAD calls consistent by specifying packages as strings
(previously it was a mix of strings and keywords)
- wrapped the infinite loop inside BT:MAKE-THREAD, that way the user can
interact with the image while it's running (i.e. call BT:ALL-THREADS to
confirm the Swank server is up and running)